added link functions and refactored file structure
