+++ pymdbapi/schema/manufacturers_schema.py
