+++ pymdbapi/schema/operating_systems_schema.py
